Decrypt repo secrets after checkout
